refactored feature tests
