+++ app/baseball/tests/feature/test_competition_api.py
@@ -49,3 +49,45 @@ class TestCompetitionDetailsApi (APITestCase):
 
         get_response = self.client.get(path=f'/api/baseball/competitions/{self.test_competition_id}/')
         self.assertEqual(status.HTTP_404_NOT_FOUND, get_response.status_code)
+
+class TestCompetitionListApi (APITestCase):
+    """Tests for endpoints defined in the CompetitionList view.
+    """
+    fixtures = ['user']
+
+    def setUp(self) -> None:
+        """Set up necessary objects for testing.
+        """
+        Competition.objects.create(
+            name = 'Test Season',
+            type = Competition.CompetitionType.SEASON
+        )
+        Competition.objects.create(
+            name = 'Test Tournament',
+            type = Competition.CompetitionType.TOURNAMENT
+        )
+        self.client = APIClient()
+        user = User.objects.get(username='DeveloperAdmin')
+        self.client.force_authenticate(user)
+        return super().setUp()
+    
+    def test_create_competition(self):
+        """Test the POST endpoint for creating a competition.
+        """
+        competition_data = {
+            'name' : 'Test Season',
+            'type' : Competition.CompetitionType.SEASON
+        }
+        response = self.client.post('/api/baseball/competitions/', data=competition_data, format='json')
+        
+        self.assertEqual(status.HTTP_201_CREATED, response.status_code)
+        self.assertEqual(competition_data.get('name'), response.data.get('name'))
+        self.assertEqual(competition_data.get('type'), response.data.get('type'))
+
+    def test_competitions_list(self):
+        """Test the GET endpoint for getting the list of competitions.
+        """
+        response = self.client.get('/api/baseball/competitions/')
+
+        self.assertEqual(status.HTTP_200_OK, response.status_code)
+        self.assertEqual(2, len(response.data))
